♻️ Extract image loader as a module
To align with `FontLoader`, this commit extracts the `ImageLoader` as a
separate module.
