+++ src/document.ts
@@ -3,7 +3,8 @@ import { PDFDict, PDFDocument, PDFHexString, PDFName } from 'pdf-lib';
 import { Size } from './box.js';
 import { createFontLoader } from './font-loader.js';
 import { createFontStore, FontStore } from './fonts.js';
-import { createImageLoader, createImageStore, ImageStore } from './images.js';
+import { createImageLoader } from './image-loader.js';
+import { createImageStore, ImageStore } from './images.js';
 import { Page } from './page.js';
 import { applyOrientation, paperSizes } from './page-sizes.js';
 import { DocumentDefinition, Metadata } from './read-document.js';
